pyproject.toml
[project]
name = "copilot-key-handler"
version = "0.1.0"
description = "Maps the Windows Copilot key to launch Claude Code on Linux"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "evdev>=1.7.0",
    "pyyaml>=5.0",
]
